# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Retrieve a single news article",
                "description": "Fetch a specific news article using its unique ID from the cache.",
                "operationId": "89e0d2121875c33ae566a67b105a3718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the news article to retrieve.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the news article.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Breaking News: Local Event"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Details about the event..."
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "guardian"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-20T14:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News not found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/feed": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Retrieve the personalized news feed for the authenticated user",
                "description": "Fetch a personalized news feed based on the user's preferences (sources and categories), with pagination support.",
                "operationId": "33188e71bc3bec6d22ad9e6e61984f89",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number to fetch. Defaults to 1.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the personalized news feed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Breaking News: Local Event"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Details about the event..."
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "source": {
                                                "type": "string",
                                                "example": "guardian"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-20T14:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, no valid authentication token provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No news found for the given user and preferences.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No news found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer token for authentication. Example: \"Authorization: Bearer {token}\""
            }
        }
    }
}